refactor(Message): rename fron alert to message
