+++ src/Message/README.md
@@ -1,9 +1,9 @@
-# Alert
+# Message
 
 ### Usage
 
 ```jsx
-import { Alert } from '@tillersystems/stardust';
+import { Message } from '@tillersystems/stardust';
 ```
 
 <!-- STORY -->
@@ -11,15 +11,13 @@ import { Alert } from '@tillersystems/stardust';
 ### Properties
 
 - `className` - className needed by styled components.
-- `closable` - whether it is possible to close the alert box.
-- `message` - An alert can have a message description.
+- `description` - An alert can have a message description.
 - `onClose` -An alert can have a clickable button to close it.
 - `type` - The type of the message box.
 
 | propName    |                       propType                       | defaultValue | isRequired |
 | ----------- | :--------------------------------------------------: | :----------: | :--------: |
 | `className` |                       `string`                       |     `''`     |            |
-| `closable`  |                        `bool`                        |   `false`    |            |
 | `message`   |                       `string`                       |              |     \*     |
 | `onClose`   |                        `func`                        |  `() => {}`  |            |
 | `type`      | `enum: 'info', 'success', 'warning', 'error' 'info'` |    `info`    |            |
@@ -27,11 +25,11 @@ import { Alert } from '@tillersystems/stardust';
 ### Example
 
 ```jsx
-import { Alert } from '@tillersystems/stardust';
+import { Message } from '@tillersystems/stardust';
 
 render() {
   return (
-    <Alert message="this is a message" type="success">
+    <Message description="this is a message" type="success">
   );
 }
 ```
