+++ src/Message/__tests__/Message.test.js
@@ -1,41 +1,35 @@
 import React from 'react';
 import 'jest-styled-components';
 
-import Alert from '..';
+import Message from '..';
 
-describe('<Alert />', () => {
+describe('<Message />', () => {
   it('should render without a problem', () => {
-    const render = shallowWithTheme(<Alert message="this is a message" />);
-
-    expect(render.dive()).toMatchSnapshot();
-  });
-
-  it('should be closable', () => {
-    const render = shallowWithTheme(<Alert message="this is a message" closable />);
+    const render = shallowWithTheme(<Message description="this is a message" />);
 
     expect(render.dive()).toMatchSnapshot();
   });
 
   it('should render a success alert', () => {
-    const render = shallowWithTheme(<Alert message="this is a message" type="success" />);
+    const render = shallowWithTheme(<Message description="this is a message" type="success" />);
 
     expect(render.dive()).toMatchSnapshot();
   });
 
   it('should render a info alert', () => {
-    const render = shallowWithTheme(<Alert message="this is a message" type="info" />);
+    const render = shallowWithTheme(<Message description="this is a message" type="info" />);
 
     expect(render.dive()).toMatchSnapshot();
   });
 
   it('should render a warning alert', () => {
-    const render = shallowWithTheme(<Alert message="this is a message" type="warning" />);
+    const render = shallowWithTheme(<Message description="this is a message" type="warning" />);
 
     expect(render.dive()).toMatchSnapshot();
   });
 
   it('should render a error alert', () => {
-    const render = shallowWithTheme(<Alert message="this is a message" type="error" />);
+    const render = shallowWithTheme(<Message description="this is a message" type="error" />);
 
     expect(render.dive()).toMatchSnapshot();
   });
@@ -43,7 +37,7 @@ describe('<Alert />', () => {
   it('should call onClose handler', () => {
     const handleOnClose = jest.fn();
     const render = mountWithTheme(
-      <Alert message="this is a message" closable onClose={handleOnClose} />,
+      <Message description="this is a message" onClose={handleOnClose} />,
     );
 
     render
