+++ src/Message/elements.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Description = styled.span`
+  flex: 1;
+  text-align: left;
+  margin-left: 2rem;
+`;
