+++ src/Message/helpers.js
@@ -1,0 +1,8 @@
+export const getStatusBackgroundColor = ({ palette }, type) => {
+  return {
+    info: palette.primary.default,
+    success: palette.success.default,
+    error: palette.warning.default,
+    warning: palette.failure.default,
+  }[type];
+};
